% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muse_strip_names.R
\name{muse_diagnoses}
\alias{muse_diagnoses}
\title{Read all diagnoses in an XML file}
\usage{
muse_diagnoses(file)
}
\arguments{
\item{file}{Path to the input XML file}
}
\description{
Generate a data frame of all values from `StmtFlag` and `StmtText` nodes
    within the `OriginalDiagnosis` and `Diagnosis` sections of an XML file.
}
\examples{
file <- muse_example("muse/muse_ecg1.xml")
diagnoses <- muse_diagnoses(file)
dplyr::count(diagnoses, value, sort = TRUE)
}
