% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muse_strip_names.R
\name{muse_example}
\alias{muse_example}
\title{Get path to sample XML file}
\usage{
muse_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If NULL, the example files will be listed.}
}
\description{
musePHI comes with a few sample files in the inst/extdata directory.
    This is a convenience function to access them
}
\examples{
muse_example(path = NULL)
muse_example(path = "muse/muse_ecg1.xml")
}
