% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muse_strip_names.R
\name{muse_deidentify}
\alias{muse_deidentify}
\title{De-identify Muse XML Files}
\usage{
muse_deidentify(file, output_file, replace)
}
\arguments{
\item{file}{A string specifying the path to the input XML file to be de-identified.}

\item{output_file}{A string specifying the path where the de-identified XML file will be saved.}

\item{replace}{A named list where each name is a path to an XML node (relative to the `RestingECG` root) and its value is the replacement text.
For example, `list("PatientDemographics/PatientLastName" = "LastName")` would replace the text of the `PatientLastName` element within `PatientDemographics`.}
}
\value{
Invisibly returns the path to the output file containing the de-identified XML document. The function primarily operates through side effects (reading an input file, modifying its content, and writing the result to a new file).
}
\description{
This function de-identifies XML files containing patient data from Muse ECG systems by replacing specified elements with anonymized values.
The function targets specific nodes within the `RestingECG` element of the XML document based on a provided list of elements to be replaced.
}
\note{
Make sure the paths provided in `replace` accurately reflect the structure of your XML documents.
}
\examples{
# De-identify a sample Muse ECG XML file
replace <- list(
  "PatientDemographics/PatientLastName" = "LastName",
  "PatientDemographics/PatientFirstName" = "FirstName",
  "PatientDemographics/PatientID" = "AnonymousID",
  "TestDemographics/AcquisitionDate" = "01-01-0001",
  "TestDemographics/AcquisitionTime" = "00:00:00"
)
file <- muse_example("muse/muse_ecg1.xml")
output_file <- tempfile(fileext = ".xml")
muse_deidentify(file, output_file, replace)

}
\seealso{
\code{\link[xml2]{read_xml}}, \code{\link[xml2]{write_xml}} for the underlying XML manipulation functions used.
}
